Bsls atomic signed twos compliment drqs 175695137 (#4829)
* bsls_atomicoperations: make non-assembly code not overflow signed integral values

+++ groups/bsl/bsls/bsls_atomicoperations_x86_win_msvc.h
@@ -193,10 +193,11 @@ inline
 int AtomicOperations_X86_WIN_MSVC::
     addIntNv(AtomicTypes::Int *atomicInt, int value)
 {
-    return _InterlockedExchangeAdd(
-            reinterpret_cast<long volatile *>(&atomicInt->d_value),
-            value)
-        + value;
+    return static_cast<unsigned>(_InterlockedExchangeAdd(
+                                 reinterpret_cast<long volatile *>(
+                                                          &atomicInt->d_value),
+                                 value))
+           + value;
 }
 
     // *** atomic functions for Int64 ***
